Move Widget.cpp to libopenrct2ui
