+++ src/openrct2/Game.cpp
@@ -28,7 +28,6 @@
 #include "Input.h"
 #include "interface/Screenshot.h"
 #include "interface/Viewport.h"
-#include "interface/Widget.h"
 #include "interface/Window.h"
 #include "localisation/Localisation.h"
 #include "management/Finance.h"
